add teuthology-dispatcher command
This commit does the following:

1. Add a new dispatcher command, teuthology-dispatcher, that runs
each job in the queue as its subprocess by invoking
teuthology-worker.

2. Modify teuthology-worker such that it only runs the job whose
config is provided to it.
