+++ scripts/dispatcher.py
@@ -1,0 +1,37 @@
+import argparse
+
+import teuthology.dispatcher
+
+
+def main():
+    teuthology.dispatcher.main(parse_args())
+
+
+def parse_args():
+    parser = argparse.ArgumentParser(description="""
+Start a dispatcher for the specified tube. Grab jobs from a beanstalk
+queue and run the teuthology tests they describe as subprocesses. One
+job is run at a time.
+""")
+    parser.add_argument(
+        '-v', '--verbose',
+        action='store_true', default=None,
+        help='be more verbose',
+    )
+    parser.add_argument(
+        '--archive-dir',
+        metavar='DIR',
+        help='path under which to archive results',
+        required=True,
+    )
+    parser.add_argument(
+        '-l', '--log-dir',
+        help='path in which to store logs',
+        required=True,
+    )
+    parser.add_argument(
+        '-t', '--tube',
+        help='which beanstalk tube to read jobs from',
+        required=True,
+    )
+    return parser.parse_args()
