+++ setup.py
@@ -129,6 +129,7 @@ setup(
             'teuthology-prune-logs = scripts.prune_logs:main',
             'teuthology-describe = scripts.describe:main',
             'teuthology-reimage = scripts.reimage:main'
+            'teuthology-dispatcher = scripts.dispatcher:main'
             ],
         },
